mutualize getFirstAndLastBlocks
